backend/pyproject.toml
[project]
name = "document-intelligence"
version = "0.1.0"
description = "Document Intelligence Backend"
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.13,<3.14"
dependencies = [
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.4",
    "uvicorn>=0.32.0",
    "boto3>=1.35.70",
    "botocore>=1.35.70",
    "voyageai>=0.3.2",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0",
    "pdf2image>=1.16.3",
    "python-docx>=1.1.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-mongodb>=0.2.0",
    "langchain-aws>=0.2.0",
    "langchain-voyageai>=0.1.0",
    "requests>=2.31.0",
    "langgraph-store-mongodb>=0.1.0",
    "langgraph-checkpoint-mongodb>=0.1.0",
    "schedule>=1.2.0",
    "matplotlib>=3.10.6",
    "reportlab>=4.4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
